chore: Remove Makefile powered test suite
While still considered to be an impressive technical achievement, it
became a burden to maintain the test suite powered by `make` and backed
by Docker. Most of what it provided is provided by Github too and the
only maintainer has all the minor details in muscle memory.

Long story short: Developing this project is easier and more fun without
this local-only test suite.